Increase configurability
I should really start accepting command line flags…

+++ scrapper.py
@@ -18,18 +18,23 @@ from progress.bar import Bar
 from lib.giustizia import get_case_details
 from lib.ranges import load_ids_from_json
 
+READ_ONLY = True
+RANGE_ONLY = True
+RATE_LIMIT = 0.1
+
 INITIAL_ID = 0
 FINAL_ID = 60000
 RANGE_YEAR = 2018
 
 query_range = {RANGE_YEAR: range(INITIAL_ID, FINAL_ID, 1)}
 
-if os.path.exists("json_results.txt"):
+if os.path.exists("json_results.txt") and not RANGE_ONLY:
     loaded_json = load_ids_from_json("json_results.txt")
     query_range = loaded_json if loaded_json else query_range
 
-json_results = open("json_results.txt", "w+")
-csv_results = open("csv_results.csv", "w+")
+if not READ_ONLY:
+    json_results = open("json_results.txt", "w+")
+    csv_results = open("csv_results.csv", "w+")
 
 for year in query_range:
     print("Querying cases from year {}".format(year))
@@ -39,14 +44,16 @@ for year in query_range:
 
         if case:
             print(" - {}".format(case))
-            json_results.write(json.dumps(case.asdict()))
-            csv_results.write(str(case))
-            json_results.write("\n")
-            csv_results.write("\n")
-            json_results.flush()
-            csv_results.flush()
+            if not READ_ONLY:
+                json_results.write(json.dumps(case.asdict()))
+                csv_results.write(str(case))
+                json_results.write("\n")
+                csv_results.write("\n")
+                json_results.flush()
+                csv_results.flush()
 
-        time.sleep(0.1)  # wait a little to prevent DOS
+        time.sleep(RATE_LIMIT)  # wait a little to prevent DOS
 
-json_results.close()
-csv_results.close()
+if not READ_ONLY:
+    json_results.close()
+    csv_results.close()
